refactor(api): move route
to identity-access-management bounded context
