+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.route.js
@@ -106,4 +106,30 @@ export const oidcProviderRoutes = [
       tags: ['identity-access-management', 'api', 'oidc'],
     },
   },
+  {
+    method: 'POST',
+    path: '/api/oidc/user/check-reconciliation',
+    config: {
+      auth: false,
+      validate: {
+        payload: Joi.object({
+          data: Joi.object({
+            attributes: Joi.object({
+              email: Joi.string().email().required(),
+              password: Joi.string().required(),
+              'identity-provider': Joi.string().required(),
+              'authentication-key': Joi.string().required(),
+            }),
+            type: Joi.string(),
+          }),
+        }),
+      },
+      handler: (request, h) => oidcProviderController.findUserForReconciliation(request, h),
+      notes: [
+        "- Cette route permet d'identifier un utilisateur Pix provenant de la double mire OIDC.\n" +
+          "- Elle renvoie un objet contenant des informations sur l'utilisateur.",
+      ],
+      tags: ['identity-access-management', 'api', 'oidc'],
+    },
+  },
 ];
