--- api/tests/integration/application/authentication/oidc/index_test.js
@@ -1,95 +1,0 @@
-import * as moduleUnderTest from '../../../../../lib/application/authentication/oidc/index.js';
-import { oidcController } from '../../../../../lib/application/authentication/oidc/oidc-controller.js';
-import {
-  AuthenticationKeyExpired,
-  DifferentExternalIdentifierError,
-  UserNotFoundError,
-} from '../../../../../lib/domain/errors.js';
-import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
-
-describe('Integration | Application | Route | OidcRouter', function () {
-  describe('POST /api/oidc/user/check-reconciliation', function () {
-    context('error cases', function () {
-      context('when user is not found', function () {
-        it('should return a response with HTTP status code 404', async function () {
-          // given
-          sinon.stub(oidcController, 'findUserForReconciliation').rejects(new UserNotFoundError());
-          const httpTestServer = new HttpTestServer();
-          httpTestServer.setupAuthentication();
-          await httpTestServer.register(moduleUnderTest);
-
-          const response = await httpTestServer.requestObject({
-            method: 'POST',
-            url: '/api/oidc/user/check-reconciliation',
-            payload: {
-              data: {
-                attributes: {
-                  email: 'eva.poree@example.net',
-                  password: 'pix123',
-                  'identity-provider': 'POLE_EMPLOI',
-                  'authentication-key': '123abc',
-                },
-              },
-            },
-          });
-
-          // then
-          expect(response.statusCode).to.equal(404);
-          expect(response.result.errors[0].detail).to.equal('Ce compte est introuvable.');
-        });
-      });
-
-      context('when authentication key expired', function () {
-        it('should return a response with HTTP status code 401', async function () {
-          // given
-          sinon.stub(oidcController, 'findUserForReconciliation').rejects(new AuthenticationKeyExpired());
-          const httpTestServer = new HttpTestServer();
-          httpTestServer.setupAuthentication();
-          await httpTestServer.register(moduleUnderTest);
-
-          const response = await httpTestServer.request('POST', `/api/oidc/user/check-reconciliation`, {
-            data: {
-              attributes: {
-                email: 'eva.poree@example.net',
-                password: 'pix123',
-                'identity-provider': 'POLE_EMPLOI',
-                'authentication-key': '123abc',
-              },
-            },
-          });
-
-          // then
-          expect(response.statusCode).to.equal(401);
-          expect(response.result.errors[0].detail).to.equal('This authentication key has expired.');
-        });
-      });
-
-      context('when external identity id and external identifier are different', function () {
-        it('should return a response with HTTP status code 412', async function () {
-          // given
-          sinon.stub(oidcController, 'findUserForReconciliation').rejects(new DifferentExternalIdentifierError());
-          const httpTestServer = new HttpTestServer();
-          httpTestServer.setupAuthentication();
-          await httpTestServer.register(moduleUnderTest);
-
-          const response = await httpTestServer.request('POST', `/api/oidc/user/check-reconciliation`, {
-            data: {
-              attributes: {
-                email: 'eva.poree@example.net',
-                password: 'pix123',
-                'identity-provider': 'POLE_EMPLOI',
-                'authentication-key': '123abc',
-              },
-            },
-          });
-
-          // then
-          expect(response.statusCode).to.equal(409);
-          expect(response.result.errors[0].detail).to.equal(
-            "La valeur de l'externalIdentifier de la méthode de connexion ne correspond pas à celui reçu par le partenaire.",
-          );
-        });
-      });
-    });
-  });
-});
